Updated readme, added requirements, cleaned database.py code, added to it some checks for validity

+++ database.py
@@ -1,5 +1,6 @@
 import sqlite3
 from tkinter import *
+from tkinter import messagebox
 import os
 from utils.tk_table_dke  import Table
 
@@ -9,8 +10,8 @@ Create a database, connect it to program
 """
 
 # TODO: поле описания сделать расширяемым (кнопку для открытия поля "Описание" в отдельном окне?)
-# TODO: добавить запрос подтверждения на удаление (askyesno и if?)
 # TODO: запрет на добавление пустой записи в базу (предупреждение или что-то типа)
+# TODO: добавить заголовки в колонках таблицы с записями БД
 
 def database():
     """
@@ -50,48 +51,62 @@ def database():
 
     def delete():
         # Delete a chosen record from db (type in choose_box)
+        ask_yesno = messagebox.askquestion('Удалить запись',
+                                           'Вы уверены, что хотите удалить запись?',
+                                           icon = 'warning')
+        if ask_yesno == 'yes':
+            conn = sqlite3.connect('database_med.db')
+            c = conn.cursor()
 
-        conn = sqlite3.connect('database_med.db')
-        c = conn.cursor()
+            # Delete a record (by record's ID)
+            c.execute("DELETE FROM database"
+                      " WHERE oid=" + "'" + choose_box.get() + "'")
+            # Clear the choose_box
+            choose_box.delete(0, END)
 
-        # Delete a record (by record's ID)
-        c.execute("DELETE FROM database"
-                  " WHERE oid=" + "'" + choose_box.get() + "'")
-        # Clear the choose_box
-        choose_box.delete(0, END)
-
-        # Commit changes and close connection
-        conn.commit()
-        conn.close()
+            # Commit changes and close connection
+            conn.commit()
+            conn.close()
+        else: pass
 
     def submit():
         # Confirm adding new record to the db
 
-        conn = sqlite3.connect('database_med.db')
-        c = conn.cursor()
+        # Prevent adding empty record to the DB
+        for entry in (topic.get(),
+                      description.get("1.0", "end-1c")):
+            if len(entry) == 0:
+                messagebox.showerror(title="Пустая запись не допустима",
+                                     message="Поля 'Тема' и 'Описание' "
+                                             "должны быть заполнены")
+                break
+            else:
+                conn = sqlite3.connect('database_med.db')
+                c = conn.cursor()
 
-        c.execute("INSERT INTO database VALUES (:topic, :description, :symptoms, :treatment, :picts_videos, :notes)",
-                  {
-                      'topic': topic.get(),
-                      'description': description.get("1.0", "end-1c"),
-                      'symptoms': symptoms.get(),
-                      'treatment': treatment.get(),
-                      'picts_videos': picts_videos.get(),
-                      'notes': notes.get()
-                  }
-                  )
+                c.execute("INSERT INTO database VALUES (:topic, :description, :symptoms, :treatment, :picts_videos, :notes)",
+                          {
+                              'topic': topic.get(),
+                              'description': description.get("1.0", "end-1c"),
+                              'symptoms': symptoms.get(),
+                              'treatment': treatment.get(),
+                              'picts_videos': picts_videos.get(),
+                              'notes': notes.get()
+                          }
+                          )
 
-        # Commit changes and close connection
-        conn.commit()
-        conn.close()
+                # Commit changes and close connection
+                conn.commit()
+                conn.close()
 
-        # Clear the textboxes
-        topic.delete(0, END)
-        description.delete("1.0", END)
-        symptoms.delete(0, END)
-        treatment.delete(0, END)
-        picts_videos.delete(0, END)
-        notes.delete(0, END)
+                # Clear the textboxes
+                topic.delete(0, END)
+                description.delete("1.0", END)
+                symptoms.delete(0, END)
+                treatment.delete(0, END)
+                picts_videos.delete(0, END)
+                notes.delete(0, END)
+                break
 
     def query():
         """
@@ -106,6 +121,7 @@ def database():
         c.execute('SELECT oid, * FROM database')
         records = c.fetchall()
 
+        # Display records in the form of a table
         record_list = Toplevel()
         Table(records, record_list)
 
@@ -113,15 +129,14 @@ def database():
         conn.commit()
         conn.close()
 
-    # Create Edit function to update record
     def save_changes():
-        # Create a database or connect to one
+        # Commit changes to the chosen db record
+
         conn = sqlite3.connect('database_med.db')
-        # Create cursor
         c = conn.cursor()
 
-        c.execute("""
-            UPDATE database SET
+        # Update record, chosen by its ID (oid)
+        c.execute("""UPDATE database SET
             topic = :topic,
             description = :description,
             symptoms = :symptoms,
@@ -131,8 +146,9 @@ def database():
             
             WHERE oid = :oid""",
             {
+                "oid": choose_box.get(),
                 "topic": topic_editor.get(),
-                "description": description_editor.get(),
+                "description": description_editor.get("1.0", END),
                 "symptoms": symptoms_editor.get(),
                 "treatment": treatment_editor.get(),
                 "picts_videos": picts_videos_editor.get(),
@@ -140,21 +156,20 @@ def database():
             }
         )
 
-        # Commit changes
+        # Commit changes and close connection
         conn.commit()
-        # Close connection
         conn.close()
 
     def edit():
-        # Creating tkinter main window
+        # Opens a GUI for editing db record, chosen in the choose_box
+
+        # Creating a new window
         editor = Tk()
         editor.title('РЕДАКТИРОВАНИЕ ЗАПИСИ')
         editor.geometry('620x400')
         editor.iconbitmap('squirrel.ico')
 
-        # Create a database or connect to one
         conn = sqlite3.connect('database_med.db')
-        # Create cursor
         c = conn.cursor()
 
         # Query the database (by record's ID)
@@ -162,7 +177,7 @@ def database():
         c.execute('SELECT * FROM database WHERE oid=' + '"' + record_id + '"')
         records = c.fetchall()
 
-        # Create global variables for text box names
+        # Make edit boxes names global
         global topic_editor
         global description_editor
         global symptoms_editor
@@ -198,7 +213,7 @@ def database():
         notes_label_editor = Label(editor, text='Примечание', width=15, anchor=E)
         notes_label_editor.grid(row=5, column=0)
 
-        # Loop through results
+        # Insert record's info into entry boxes of the Edit window
         for record in records:
             topic_editor.insert(0, record[0])
             description_editor.insert("1.0", record[1])
@@ -207,8 +222,8 @@ def database():
             picts_videos_editor.insert(0, record[4])
             notes_editor.insert(0, record[5])
 
-        # Create a Clear All Fields function for Edit window
         def clear_all_editor():
+            # Clear all fields in the Edit window
             topic_editor.delete(0, END)
             description_editor.delete("1.0", END)
             symptoms_editor.delete(0, END)
@@ -224,11 +239,9 @@ def database():
         clear_all = Button(editor, text='Очистить все поля', command=clear_all_editor)
         clear_all.grid(row=7, column=0, columnspan=2, pady=5, padx=10, ipadx=110)
 
-        # Commit changes
+        # Commit changes and close connection
         conn.commit()
-        # Close connection
         conn.close()
-
 
 
     # Create entry boxes
@@ -282,12 +295,9 @@ def database():
     edit_btn = Button(root, text='Редактировать запись\n(по названию темы)', command=edit)
     edit_btn.grid(row=10, column=1, pady=10, padx=(80,10))
 
-    # Commit changes
+    # Commit changes and close connection
     conn.commit()
-
-    # Close connection
     conn.close()
-
 
     root.mainloop()
